requirements.txt
# Financial Research Tool Demo - Dependencies

# Core Framework
streamlit==1.28.0

# Data Acquisition
yfinance==0.2.31
newsapi-python==0.2.7
requests==2.31.0
beautifulsoup4==4.12.2

# Data Processing & Analysis
pandas==2.1.3
numpy==1.26.2

# Natural Language Processing (for sentiment analysis)
nltk==3.8.1
textblob==0.17.1
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99

# Data Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Database
# SQLite3 is built into Python, no separate installation needed

# Utilities
python-dateutil==2.8.2
pytz==2023.3


polygon-api-client==1.13.2
python-dotenv==1.0.0
